Hibernate is configured and working.
Added endpoint to create bookings.

Added a DB constraint to make the dates unique.

Added implementation for getAvailableDates.

Fixed a bug that did not allow the user to book the campsite for a single day (same arrival and departure dates). Added triggers to block booking if the dates are not available

Added endpoint to cancel bookings.

Added endpoint to update bookings.

Refactored project structure. Added a DAO.

Exceptions are thrown in the Service and and in the DAO so that the
Controller could catch them and return a response with the appropriate
http status code.

Bookings cannot be created between a reserved date.

Cancel and Update cannot be performed on bookings that do not exist.